initial modeling and template structuring
