+++ linkml/generators/pydanticgen/__init__.py
@@ -1,0 +1,1 @@
+from linkml.generators.pydanticgen.pydanticgen import PydanticGenerator
